pyproject.toml
[tool.poetry]
name = "complaintsystemmaincourseapp"
version = "0.1.0"
description = ""
authors = ["Praveen Soman <praveens79@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.92.0"
uvicorn = "^0.20.0"
asyncpg = "^0.27.0"
psycopg2-binary = "^2.9.5"
databases = "^0.7.0"
alembic = "^1.9.4"
python-decouple = "^3.7"
pyjwt = "^2.6.0"
passlib = "^1.7.4"
asyncclick = "^8.1.3.4"
boto3 = "^1.26.103"
requests = "^2.28.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
